# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant Chain POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff
      parameters:
        - name: outletId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: Not found
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Menu Categories ───────────────────────────────────────────────────────
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List tables for an outlet
      parameters:
        - name: outletId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Orders ────────────────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Staff:
      type: object
      required: [id, outletId, name, role, createdAt]
      properties:
        id:
          type: integer
        outletId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
          enum: [super_admin, manager, cashier, kitchen]
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, role, pin]
      properties:
        outletId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
          enum: [super_admin, manager, cashier, kitchen]
        pin:
          type: string
    StaffUpdate:
      type: object
      properties:
        outletId:
          type: ["integer", "null"]
        name:
          type: string
        role:
          type: string
          enum: [super_admin, manager, cashier, kitchen]
        pin:
          type: string

    # ── Menu schemas ──────────────────────────────────────────────────────
...
    Area:
      type: object
      required: [id, outletId, name, type, color]
      properties:
        id:
          type: integer
        outletId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [standard, timed]
        hourlyRate:
          type: ["number", "null"]
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Table:
      type: object
      required: [id, outletId, name, capacity, status]
      properties:
        id:
          type: integer
        outletId:
          type: integer
        areaId:
          type: ["integer", "null"]
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, bill_requested]
        area:
          $ref: "#/components/schemas/Area"
    TableInput:
      type: object
      required: [outletId, name, capacity]
      properties:
        outletId:
          type: integer
        areaId:
          type: ["integer", "null"]
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, bill_requested]
    TableUpdate:
      type: object
      properties:
        areaId:
          type: ["integer", "null"]
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, bill_requested]

    # ── Order schemas ─────────────────────────────────────────────────────
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string

    # ── Report schemas ────────────────────────────────────────────────────
...